Decentralized VPN: Mysterium Network
Network stats:
- https://testnet.mysterium.network/

FAQ:
- https://mysterium.network/general-questions/

Source code:
- https://github.com/Mysterium

Normally I do not add any service which claims to be "Hack-resistant" even if it would cure cancer, because that is ridiculous. However, layered protection (multiple line of defenses) are in place and this makes it "harder" to hack the service. It's FOSS and this makes it easier to review/audit.