Added Strixx to custom google search engines
Strixx is a tool I made as an answer to filechef being made using jQuery and bootstrap. I thought better can be done using everything as vanilla as possible.

It is a PWA, has Dark mode, clean UI, more options, basically a successor to filechef.

Keep in mind though that the UI is heavily under development but it will work effortlessly nevertheless.