De-List all none and possible more insecure websites including all popular ones
This is basically a "political decision" and I do not make exceptions .. except when I do!

There are 23 mentions of HTTP in the Readme.md and they stay until I found a reliable alternatives, the reason why they are still on the list is that there are no secure alternatives or no alternatives at all.

Why de-list HTTP pages in the first place?
I keep it short:
- Quality over quantity (we have enough secure alternatives [in most cases])
- Speed reasons, modern pages and encryption methods loading faster. 
- Security reasons
- You get a certificate for free e.g. via Let's Encrypt, so think about it.. Why is someone not going to secure his own domain/website? If you cannot even secure your page then please think about the quality of their releases and their true intention why the website was created in the first place (maybe just to get clicks, fame, fake website etc). There are exceptions like libgen and firtgirl (only some examples), they are trusted by a lot of people but still getting de-listed. I believe pointing the finger on them might force them to change their website and call them out. People might now start asking the right question.

The cleanup will continue over the following weeks, the goal would be to only list secure, open source & reliable + trusted websites/tools & scripts from more or less trusted websites that are monitored by lots of people.